Collector - Rework finished: 1. Log processing improved thanks to sampling. 2. Fixed duplicite processing of last log file after re/start. 3. Improved support for external sql normalization - any process can be executed.

+++ IndexSuggestions.Collector.Contracts/QueryPlanNode.cs
@@ -8,7 +8,7 @@ namespace IndexSuggestions.Collector.Contracts
     {
         public decimal StartupCost { get; set; }
         public decimal TotalCost { get; set; }
-        public long? IndexId { get; set; }
+        public uint? IndexId { get; set; }
 
         public List<QueryPlanNode> Plans { get; private set; }
